Typescript and docker fixes
