+++ backend/Dockerfile
@@ -1,12 +1,21 @@
-FROM node:21-alpine
-
+FROM node:21-alpine AS builder
 WORKDIR /app
 
 COPY package*.json ./
-RUN npm install --production
+RUN npm install
 
-COPY . .
+COPY tsconfig.json ./
+COPY src ./src
 
-CMD ["npm", "start"]`
+RUN npm run build
+
+FROM node:21-alpine
+WORKDIR /app
+
+COPY package*.json ./
+RUN npm install
+
+COPY --from=builder /app/dist ./dist 
 
 EXPOSE 3000
+CMD ["node", "dist/index.js"]
