+++ backend/src/controllers/authController.ts
@@ -1,0 +1,9 @@
+import { Request, Response } from "express"
+
+export async function registerUser(req: Request, res: Response) {
+  res.send({message: "Miaw!"})
+}
+
+export async function loginUser(req: Request, res: Response) {
+  res.send({message: "Miaw miaw!!!!"})
+}
