+++ docker-compose.yaml
@@ -1,5 +1,3 @@
-version: "3.9"
-
 services:
   db:
     image: postgres:16-alpine
@@ -39,7 +37,7 @@ services:
     depends_on:
       - backend
     ports:
-      - "8080:5173"
+      - "5173:5173"
     volumes:
       - ./frontend:/app
       - /app/node_modules
