+++ frontend/Dockerfile
@@ -1,9 +1,10 @@
-FROM node:20-alpine
+FROM node:21-alpine
 
 WORKDIR /app
 
 COPY package*.json ./
-RUN npm install
+
+RUN npm install --include=dev
 
 COPY . .
