+++ frontend/package.json
@@ -35,6 +35,7 @@
     "path": "^0.12.7",
     "prettier": "3.6.2",
     "sass": "^1.92.1",
+    "sass-embedded": "^1.92.1",
     "typescript": "~5.8.0",
     "vite": "^7.0.6",
     "vite-plugin-vue-devtools": "^8.0.0",
